+++ tests/feature-store/test_model.py
@@ -11,10 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from datetime import timezone, datetime, timedelta
+
 import pytest
 
 import mlrun
 from mlrun.data_types.data_types import ValueType
+from mlrun.datastore import ParquetSource
 from mlrun.feature_store import Entity, Feature, FeatureSet
 from mlrun.feature_store.common import parse_feature_string
 from mlrun.model import TargetPathObject
@@ -86,3 +89,53 @@ def test_different_target_path_scenarios_for_run_id(
         ) == pass_run_id
         assert mlrun.model.RUN_ID_PLACE_HOLDER not in tp_obj.get_absolute_path()
         assert tp_obj.get_absolute_path() == expected_path
+
+@pytest.mark.parametrize(
+    "time_for_source, is_through_init, should_succeed, time_delta",
+    [
+        (None, None, True, None),
+        (datetime(2021, 5, 25, 10, 30, 29, 592000), False, True, None),
+        (datetime(2021, 5, 25, 10, 30, 29, 592000), True, True, None),
+        (datetime(2021, 5, 25, 10, 30, 29, 592000, timezone.utc), False, True, 0),
+        (datetime(2021, 5, 25, 10, 30, 29, 592000, timezone.utc), True, True, 0),
+        ("2021-05-25T10:30:29.abc", False, False, None),
+        ("2021-05-25T10:30:29.abc", True, False, None),
+        ("2021-05-25T10:30:29.592", False, True, None),
+        ("2021-05-25T10:30:29.592", True, True, None),
+        ("2021-05-25T10:30:29.592Z", False, True, 0),
+        ("2021-05-25T10:30:29.592Z", True, True, 0),
+        ("2021-05-25T10:30:29.592+00:00", False, True, 0),
+        ("2021-05-25T10:30:29.592+00:00", True, True, 0),
+        ("2021-05-25T10:30:29.592+02:00", False, True, timedelta(hours=2)),
+        ("2021-05-25T10:30:29.592+02:00", True, True, timedelta(hours=2)),
+    ],
+)
+def test_parquet_source_with_iso_start_or_end_time(
+    time_for_source, is_through_init, should_succeed, time_delta
+):
+    def _test_parquet_source_with_iso_start_or_end_time(time_for_source, is_through_init, time_delta):
+        if time_delta is None:
+            tzinfo = None
+        else:
+            tzinfo = timezone.utc if time_delta == 0 else timezone(time_delta)
+        actual = datetime(2021, 5, 25, 10, 30, 29, 592000, tzinfo)
+        source = ParquetSource(
+            "srcpar",
+            path="tmp/doesnt/matter",
+            start_time=time_for_source if is_through_init else None,
+            end_time=time_for_source if is_through_init else None,
+        )
+        if not is_through_init:
+            source.start_time = time_for_source
+            source.end_time = time_for_source
+
+        if source.start_time:
+            assert source.start_time == actual
+        if source.end_time:
+            assert source.end_time == actual
+
+    if should_succeed:
+        _test_parquet_source_with_iso_start_or_end_time(time_for_source, is_through_init, time_delta)
+    else:
+        with pytest.raises(ValueError, match=r".*Invalid isoformat string:.*"):
+            _test_parquet_source_with_iso_start_or_end_time(time_for_source, is_through_init, time_delta)
